Primer avance DTOs

+++ ProyectoVentasMultinivel/src/main/java/com/udistrital/app/repository/RegionRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.udistrital.app.entity.Region;
 
-public interface RegionRepository extends JpaRepository<Region, Integer>{
+public interface RegionRepository extends JpaRepository<Region, Long>{
 
+	
 }
